refactor: 스크롤/읽음 패치 로직을 useChatScroll 훅으로 분리
